pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shapcrn"
version = "0.1.0"
description = "SBML reaction-network analysis with simulation, perturbation, and Shapley-style workflows."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy",
  "pandas",
  "scipy",
  "scikit-learn",
  "python-libsbml",
  "libroadrunner",
  "SALib",
  "matplotlib",
  "seaborn",
  "plotly",
]

[project.scripts]
shapcrn = "shapcrn.examples.mainV2:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
